Some documentation; figuring things out. See notes.
So, the accelerometer is what we need to use to figure out which way is
up. It senses acceleration due to gravity. We'll need to assume the user
is stationary when they activate the app, so we can use the initial
acceleration as the direction that gravity is in. From there, we use the
gyroscope to track changes in the direction the phone is pointing to
figure out what direction the phone is pointing now.